refactor(go-template): extract memo type predicates into memo/memo-type.ts
Move the memo type-inference predicates out of the adapter class into
pure free functions (Roadmap B): isTemplateLiteralMemo, isBooleanMemo,
isStringTernaryMemo. They read only state.moduleStringConsts and
extractPropNameFromInitialValue.

inferMemoType stays on the adapter as the orchestrator that calls the
three predicates, so no new GoEmitContext member is needed.

Output byte-identical, verified by the adapter unit (552) and
adapter-tests conformance (786) suites; tsgo + biome clean.
